backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.9
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
joblib==1.5.1
numpy==1.26.4
openai==1.95.1
pandas==2.3.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
scikit-learn==1.3.2
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.1
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.35.0

# Dependencias adicionales para la arquitectura hexagonal
aiofiles==23.2.1

# Testing (opcional)
pytest==7.4.3
pytest-asyncio==0.21.1